Adding LMDB creation notebook (With append) and changes to lmdb load to convert image to float32 array

+++ virtual/tensorDriver/driving_data.py
@@ -235,7 +235,7 @@ class HandleData:
                         info_key = key_str.split('_')
                         # Get image shape [2:None] means from index 2 to the end
                         shape_img = tuple(map(lambda x: int(x), info_key[2:None]))
-                        self.__xs.append(np.frombuffer(value, dtype=np.float32).reshape(shape_img))
+                        self.__xs.append(np.frombuffer(value, dtype=np.uint8).reshape(shape_img).astype(np.float32))
 
             self.__num_images = len(self.__xs)
 
@@ -282,4 +282,4 @@ class HandleData:
                             info_key = key_str.split('_')
                             # Get image shape [2:None] means from index 2 to the end
                             shape_img = tuple(map(lambda x: int(x), info_key[2:None]))
-                            self.__val_xs.append(np.frombuffer(value, dtype=np.float32).reshape(shape_img))
+                            self.__val_xs.append(np.frombuffer(value, dtype=np.uint8).reshape(shape_img).astype(np.float32))
